feat : [테스트 통과] +수익률, 랭킹 출력 왼료

+++ src/main/java/lotto/domain/Rank.java
@@ -46,7 +46,7 @@ public enum Rank {
     public int getMatchCount(){
         return matchCount;
     }
-    public long getPrizeMoney(){
+    public int getPrizeMoney(){
         return prizeMoney;
     }
     public boolean getMatchBonus(){
